moves items form into components folder
